move tf record serialization tests to kernel tests
